Simplify code (#18)
