+++ pkg/config/config.go
@@ -55,15 +55,15 @@ type Config struct {
 
 // Configuration values specific to the certificatee tool
 type Certificatee struct {
-	UpdateInterval  time.Duration `envconfig:"CERTIFICATEE_UPDATE_INTERVAL" default:"24h"`
+	UpdateInterval  time.Duration `envconfig:"CERTIFICATEE_UPDATE_INTERVAL" default:"10m"`
 	RenewBeforeDays int           `envconfig:"CERTIFICATEE_RENEW_BEFORE_DAYS" default:"30"`
 	// HAProxyDataPlaneAPIURLs is a comma-separated list of HAProxy Data Plane API URLs
 	// Example: "http://127.0.0.1:5555,https://haproxy2.local:5555"
-	HAProxyDataPlaneAPIURLs []string `envconfig:"HAPROXY_DATAPLANE_API_URLS" default:""`
+	HAProxyDataPlaneAPIURLs []string `envconfig:"HAPROXY_DATAPLANE_API_URLS" default:"127.0.0.1:5555"`
 	// HAProxyDataPlaneAPIUser is the username for HAProxy Data Plane API basic auth
-	HAProxyDataPlaneAPIUser string `envconfig:"HAPROXY_DATAPLANE_API_USER" default:""`
+	HAProxyDataPlaneAPIUser string `envconfig:"HAPROXY_DATAPLANE_API_USER"`
 	// HAProxyDataPlaneAPIPassword is the password for HAProxy Data Plane API basic auth
-	HAProxyDataPlaneAPIPassword string `envconfig:"HAPROXY_DATAPLANE_API_PASSWORD" default:""`
+	HAProxyDataPlaneAPIPassword string `envconfig:"HAPROXY_DATAPLANE_API_PASSWORD""`
 	// HAProxyDataPlaneAPIInsecure skips TLS certificate verification (not recommended for production)
 	HAProxyDataPlaneAPIInsecure bool `envconfig:"HAPROXY_DATAPLANE_API_INSECURE" default:"false"`
 }
