--- test/integration/dataplaneapi.yaml
@@ -1,34 +1,0 @@
-config_version: 2
-name: test-dataplaneapi
-
-dataplaneapi:
-  host: 0.0.0.0
-  port: 5555
-  scheme:
-    - http
-
-  user:
-    - name: admin
-      password: adminpwd
-      insecure: true
-
-  resources:
-    maps_dir: /tmp/haproxy-test/maps
-    ssl_certs_dir: /tmp/haproxy-certs
-    spoe_dir: /tmp/haproxy-test/spoe
-
-  transaction:
-    transaction_dir: /tmp/haproxy-test/transactions
-
-haproxy:
-  config_file: /tmp/haproxy-test/haproxy.cfg
-  haproxy_bin: haproxy
-  reload:
-    reload_delay: 5
-    reload_cmd: "kill -SIGUSR2 $(cat /tmp/haproxy-test/haproxy.pid)"
-    restart_cmd: "kill -SIGUSR2 $(cat /tmp/haproxy-test/haproxy.pid)"
-    reload_strategy: custom
-
-log:
-  log_to: stdout
-  log_level: debug
