--- test/integration/haproxy.cfg
@@ -1,34 +1,0 @@
-global
-    log stdout format raw local0 info
-    stats socket /tmp/haproxy.sock mode 660 level admin expose-fd listeners
-    stats timeout 30s
-    # SSL/TLS settings
-    ssl-default-bind-ciphers ECDHE-ECDSA-AES128-GCM-SHA256:ECDHE-RSA-AES128-GCM-SHA256
-    ssl-default-bind-options ssl-min-ver TLSv1.2
-    # Certificate storage directory
-    crt-base /tmp/haproxy-certs
-
-defaults
-    log     global
-    mode    http
-    option  httplog
-    option  dontlognull
-    timeout connect 5000
-    timeout client  50000
-    timeout server  50000
-
-frontend http_front
-    bind *:8080
-    default_backend http_back
-
-frontend https_front
-    bind *:8443 ssl crt /tmp/haproxy-certs/
-    default_backend http_back
-
-backend http_back
-    server local 127.0.0.1:9999 check
-
-# Program section for Data Plane API
-program api
-    command dataplaneapi --host 0.0.0.0 --port 5555 --haproxy-bin "$(which haproxy)" --config-file /tmp/haproxy-test/haproxy.cfg --reload-cmd "kill -SIGUSR2 1" --userlist dataplaneapi
-    no option start-on-reload
